refactor: move integration test apps to examples
