+++ .github/workflows/code-test.yml
@@ -25,8 +25,8 @@ jobs:
       - name: Unit tests
         run: bun run test
 
-      - name: End-to-end tests
-        run: bun test:e2e
+      - name: Integration tests
+        run: bun test:integration
 
       - name: Build the code
         run: bun run build
