+++ examples/di-and-lifecycle/app.module.ts
