+++ examples/di-and-lifecycle/cat/cat-internal.service.ts
