+++ examples/di-and-lifecycle/cat/cat.controller.ts
