+++ examples/di-and-lifecycle/cat/cat.module.ts
