+++ examples/di-and-lifecycle/cat/cat.service.ts
