+++ examples/di-and-lifecycle/common/common.module.ts
