+++ examples/di-and-lifecycle/common/output.service.ts
