+++ examples/di-and-lifecycle/common/storage.service.ts
