+++ examples/di-and-lifecycle/dog/dog.module.ts
