+++ examples/di-and-lifecycle/dog/dog.service.ts
