+++ examples/interceptors/cache.interceptor.ts
