+++ examples/interceptors/cat.service.ts
