+++ examples/router/command-router.ts
