+++ examples/router/output.service.ts
